fmcadc5: update project to work with new drivers
No-OS driver ad9625 was updated to use the new platform driver. This
example was updated to correctly use the new driver. The two ADCs in this
example must work in interleaving mode to offer a 5GSPS performance and
must be syncronized to work in tandem. For that purpose the i5g driver
was added. i5g driver is a driver for a custom HDL core that manages the
syncronizations of the two ADCs.

Signed-off-by: adrimbarean <Andrei.Drimbarean@analog.com>